pyproject.toml
[project]
name = "django-starter-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.3"
dependencies = [
    "channels-redis>=4.2.1",
    "channels[daphne]>=4.2.2",
    "dj-database-url>=2.3.0",
    "django>=5.2",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.18.0",
    "django-debug-toolbar>=5.1.0",
    "pre-commit>=4.1.0",
    "ruff>=0.11.0",
]
pg = [
    "psycopg[binary]>=3.2.6",
]
storages = [
    "django-storages[s3]>=1.14.6",
]
web = [
    "granian>=2.0.1",
    "redis>=5.2.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
